backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic-settings==2.7.1